Added two functions
Firstly, pad_and_encrypt that takes an array of unsigned chars, pads it and encrypts it using the key given. Returns a pointer to the encrypted message.

I also made general_decrypt (didn't know what else to call it) that decrypts an encrypted message that's longer than 16 bytes using the key given.